requirements.txt
django
django-cors-headers
djangorestframework
PyJWT
intuit-oauth==1.2.4
rauth>=0.7.3
requests>=2.26.0
simplejson>=3.17.0
six>=1.14.0
drf-yasg